+++ ui-plugins/muse-lib-react/src/features/common/polling.js
@@ -15,6 +15,7 @@ export default function({
   converter,
   retries = 30,
   stopIf,
+  onError = () => {},
 }) {
   const poller = {
     remaining: retries > 0 ? retries : 1,
@@ -27,6 +28,7 @@ export default function({
           // If re-started, discard previous response
           if (this.startTime !== timestamp) return;
           this.value = converter ? converter(res) : res;
+          // Reset remaining count
           this.remaining = retries > 0 ? retries : 1;
           this.timestamp = Date.now();
           delete this.error;
@@ -44,7 +46,7 @@ export default function({
         })
         .catch(err => {
           if (this.startTime !== timestamp) return;
-
+          onError && onError(err);
           this.error = err;
           this.errorTimestamp = Date.now();
           if (expires > 0 && this.timestamp && Date.now() - this.timestamp > expires) {