+++ ui-plugins/muse-manager/src/ext/museLayout.js
@@ -12,13 +12,31 @@ const museLayout = {
     getItems() {
       return [
         {
-          key: 'createApp',
+          key: 'createMenu',
           icon: 'PlusOutlined',
           position: 'right',
-          onClick: () => {
-            NiceModal.show('muse-manager.create-app-modal');
+          type: 'menu',
+          menuMeta: {
+            trigger: {
+              label: '+ Create',
+            },
+            items: [
+              {
+                key: 'createApp',
+                label: 'Create App',
+                onClick: () => {
+                  NiceModal.show('muse-manager.create-app-modal');
+                },
+              },
+              {
+                key: 'createPlugin',
+                onClick: () => {
+                  NiceModal.show('muse-manager.create-plugin-modal');
+                },
+                label: 'Create Plugin',
+              },
+            ],
           },
-          label: 'Create App',
         },
       ];
     },
@@ -41,12 +59,12 @@ const museLayout = {
           link: '/apps',
           label: 'Apps',
         },
-        // {
-        //   key: 'plugins',
-        //   icon: 'StarOutlined',
-        //   link: '/plugins',
-        //   label: 'Plugins',
-        // },
+        {
+          key: 'plugins',
+          icon: 'StarOutlined',
+          link: '/plugins',
+          label: 'Plugins',
+        },
       ];
     },
   },