+++ ui-plugins/muse-manager/src/hooks/index.js
@@ -1,3 +1,4 @@
 export { default as useAbility } from './useAbility';
 export { default as useSyncStatus } from './useSyncStatus';
+export { default as useSearchState } from './useSearchState';
 export * from './museHooks';