+++ workspace/packages/muse-core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ebay/muse-core
 
+## 1.0.25
+
+### Patch Changes
+
+- dev
+
+## 1.0.24
+
+### Patch Changes
+
+- Dev test.
+
 ## 1.0.21
 
 ### Patch Changes