+++ workspace/packages/muse-craco-plugin/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @ebay/muse-craco-plugin
 
+## 1.0.27
+
+### Patch Changes
+
+- dev
+- Updated dependencies
+  - @ebay/muse-dev-utils@1.0.28
+  - @ebay/muse-webpack-plugin@1.0.24
+
+## 1.0.26
+
+### Patch Changes
+
+- Dev test.
+- Updated dependencies
+  - @ebay/muse-dev-utils@1.0.26
+  - @ebay/muse-webpack-plugin@1.0.23
+
 ## 1.0.25
 
 ### Patch Changes