+++ workspace/packages/muse-dev-utils/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @ebay/muse-dev-utils
 
+## 1.0.28
+
+### Patch Changes
+
+- dev
+- Updated dependencies
+  - @ebay/muse-core@1.0.25
+  - @ebay/muse-express-middleware@1.0.25
+
+## 1.0.26
+
+### Patch Changes
+
+- Dev test.
+- Updated dependencies
+  - @ebay/muse-core@1.0.24
+  - @ebay/muse-express-middleware@1.0.23
+
 ## 1.0.25
 
 ### Patch Changes