+++ workspace/packages/muse-express-middleware/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @ebay/muse-express-middleware
 
+## 1.0.25
+
+### Patch Changes
+
+- dev
+- Updated dependencies
+  - @ebay/muse-core@1.0.25
+
+## 1.0.23
+
+### Patch Changes
+
+- Dev test.
+- Updated dependencies
+  - @ebay/muse-core@1.0.24
+
 ## 1.0.22
 
 ### Patch Changes