+++ workspace/packages/muse-plugin-acl/CHANGELOG.md
@@ -1,5 +1,21 @@
 # muse-acl-plugin
 
+## 1.0.5
+
+### Patch Changes
+
+- dev
+- Updated dependencies
+  - @ebay/muse-core@1.0.25
+
+## 1.0.4
+
+### Patch Changes
+
+- Dev test.
+- Updated dependencies
+  - @ebay/muse-core@1.0.24
+
 ## 1.0.3
 
 ### Patch Changes