+++ workspace/packages/muse-webpack-plugin/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ebay/muse-webpack-plugin
 
+## 1.0.24
+
+### Patch Changes
+
+- dev
+
+## 1.0.23
+
+### Patch Changes
+
+- Dev test.
+- Updated dependencies
+  - @ebay/muse-modules@1.0.20
+
 ## 1.0.22
 
 ### Patch Changes